Implement TextMate syntax highlighting
Didn't wanna do this too early because the language is still changing,
but at this point, it's probably done *enough* and I have enough RSL
use cases I'd actively like to write, that spending a couple of hours
getting this basic syntax highlighting working is worth it.

I didn't do too much research around this, ChatGPT recommended TextMate
as the easiest and quickest way to get something working, and it seems
to have done the trick and is supported both by Visual Studio Code and
Jetbrains (GoLand in particular), so that's ideal.

The highlighting here is definitely a bit rough and doesn't work exactly
like I want, but it's serviceable!

I include an RSL installation script for installing the extension
locally in the invokee's VSCode. Example output:

 ⚡️ Running: test -d ~/.vscode/extensions
 ⚡️ Running: test -d ./highlighting/rad-scripting-language
 ⚡️ Running: cp -r ./highlighting/rad-scripting-language ~/.vscode/extensions
 ✅ Installed RSL syntax highlighting! You may need to restart VS Code for it to take effect.