Implement v3 of the Json Field Extraction Algo
v1 was fairly naive and I hadn't thought through all the things I
wanted. v2 was closer to what I needed but it was still haphazard and
hacked together. I realized when implementing the list-with-indexing
extraction that it needed a full rewrite and up-front thinking.

I've tried to do that here. It's still not 100% thought through -- there
are some features like flattening keys across levels that I've not
thought about at this point cause I don't wanna freeze progress for too
long but I'm feeling optimistic that this V3 implementation is quite
good and will remain extendable without significant friction - it seems
sound to me.

One of the main ideas that's made a difference is how we decide how to
convert our json fields into the trie, and how we decide what a 'node'
is. The idea is that each step through the trie should be a step in the
JSON data too, or a change in scope, and thus that every node has a 1:1
mapping to any scoping of the JSON data we'll actually face. This leads
to far less hacking around and weird logic.

The other idea is to "merge" captures from children nodes when the
parent node is being evaluated, and the logic we use for merging in a
way which leads to a displayable table. TLDR:

1) If all the columns match, append rows.
2) If no columns match, but # rows does, append columns.
3) If no columns match, but # rows does *not*, but one capture has only
   only one row, then repeat the values in those columns to match the #
   rows of the other capture, and append the columns.
4) Otherwise, fail.

See code for detail on exactly what I mean here.

One regression with this commit, though you could also argue it's not
since the old behavior is also not sound, is that we now *always* wrap
captured values in an array, even if the JSON field gives no indication
that it'd result in more than one value. The reason for this is that,
even a 'single value' field may *become* an array due to the merging
logic, as a consequence of what other fields are being captured at the
same time. I really don't want fields' array/non-array status to be
dependent on the data of a particular query. You want to be certain of
your types when you write your script.

One idea is to unwrap single-value fields if *all* the fields are
single-value fields. I may do that. My hesitation though is that, as an
RSL dev is writing a script, and perhaps then adds in a non-single-value
field to a query, then now suddenly all these other single-value fields
become arrays, which might be surprising.

Maybe we need a syntax for RSL devs to assert that they want a field to
get unwrapped and that they expect it to contain a single value. Will
think on it.

Anyway, in terms of new functionality with this commit,
list-with-indexing extraction now works, and you can also have nested
list fields like json[][][1][] which you could not do before.