Implement 'range' function
Used for generating lists of numbers between some start, end, and
optionally with a step size != 1.

I intend this to be used as the way to do for-i loops as well, kinda
like how Python does it. i.e.

  for i in range(100):
    ...

will iterate from 0 to 99 (inclusive).

The implementation is naive in that we eagerly create a list on-the-spot
containing the generated elements. This means that very large numbers
can get non-performant. For example, it takes ~4-5 seconds on my
computer to run the following line:

  a = range(100000000)

That's 100 million. Realistically, I'm actually quite impressed that
this is not slower. We *might* not even need to address this in the
future, because how often do you need to generate such large numbers in
a script like RSL aims for? And if you do, how often will a few seconds
of waiting be unacceptable? I'm not opposed to optimizing this in the
future, but I worry about the implementation of that and the trade-off
between simplicity and performance here as we'll realistically need to
add some sort of "generator" type.

Back to the function itself. We somewhat unusually allow floats. For
example:

  range(0.5, 5)    -> [0.5, 1.5, 2.5, 3.5, 4.5]
  range(0, 2, 0.5) -> [0, 0.5, 1.0, 1.5]

If *any* input is a float, we generate an array of floats. If they're
all ints, then we generate an array of ints.