Implement 'confirm' function
Just a straight-forward way to get y/n input from users. I use huh for
this, though not the built-in confirm functionality it offers as, while
it's pretty, I prefer the simple y/n approach which is more traditional.
Therefore we just read a line of text from the user and if it starts
with y, then we treat that as a yes.

I might revisit the implementation and not use huh. What I don't like
about it as-is is that it deletes the prompt and input after the user
enters it, whereas I prefer when you can scroll up and see what input
you gave later.