Add 'split' function
For splitting strings on some delimiter into an array of strings.

I need to revisit this a bit when I rethink regex strings in RSL. I
don't think we can really have this "try to treat as regex, otherwise
treat as regular string" pattern since it can be surprising to users and
easily be a gotcha. I'm thinking we might want a Python-style r-string
syntax, which would involve reworking strings a bit, probably by
wrapping them in a struct and having a bool for isRegex or not, and then
accessing that where it makes sense e.g. for the split func.