rad blocks: sort columns using original data types
Previously, we'd only apply sorting *after* having stringified the rows.
This meant that, if you had a column of numbers, it'd get sorted like
this (ascending):

 10
 2
 30
 4

because the numbers were being treated like strings and thus
*alphabetically* sorted.

With this commit, we move the sorting step to right after the json field
extraction algo has run, and just prior to stringifying all the rows.
This means that all the values are still in their original types that
they were captured as with the algo. The result is that we correctly
sort e.g. *numbers* if that's what the column really is, solving the
above problem (in this case: 2, 4, 10, 30).

Since we do allow mixed types in arrays, we need to handle that i.e.
sorting an array which contains different types. There's a bit of
precedent in both Python 2 and JavaScript (I know, not great company to
keep, but I think it's acceptable for our use case). The solution is to
first sort by type. We use the following:

 bool < int/float < string < list < map

*within* each type, we sort as normal.

In the case of RSL, for now at least, I've decided to just treat all
lists and maps as equal for the sake of sorting. I'm not sure there's a
logic which would make sense for sorting them.

We intentionally sort prior to any mapping operations in the rad block.
If you really want to sort *after*, then I think you can just split the
rad block into a request block, list comprehension with mapping, and
display block. *Maybe* we could make ordering significant in rad blocks,
and so allow users to place the sort block after any mapping operations,
and have that be how to do it, but I'm not committing to ordering
mattering in rad blocks just yet, if that's indeed what we wanna do
eventually.