Add 'sort' function
I am somewhat torn between this being called 'sort' or 'sorted'. It
seems like (especially in Python) that 'sort' is for in-place and
'sorted' exists for a version that returns a new array, which is what
our RSL version here does.

However, I don't plan on having an in-place version, and if I did, we
could just add a named arg to 'sort' called 'inplace=true', for example.

I stick with 'sort' because it's more concise and the sort/sorted split
isn't *completely* universal, and because I don't have explicit plans
for an inplace sort function.

We re-use the sorting logic a previous commit added for the sorting
functionality of rad blocks.