Implement 'quiet' keyword for shell cmds to suppress announcement
By default, Rad will 'announce' (i.e. print) commands as they're
executed. Example:

  $!`ls`
  $!`echo hello`

outputs

  ⚡️ Running: ls
  pick.rsl  simple.rsl  sorting.rsl
  ⚡️ Running: echo hello
  hello

With this commit, you can now suppress that. Example:

  quiet $!`ls`
  quiet $!`echo hello`

outputs

  pick.rsl  simple.rsl  sorting.rsl
  hello

I generally like the announcement, but they can get in the way. For
example, sometimes I have scripts which end with a series of prints on
next steps or instructions after a successful script run, but if there
are deferred bash commands for cleaning things up, then those final
instructions get pushed up after the cleanups get announced, which are
not terribly interesting.

I'm not 100% sure about the 'quiet' keyword for doing this. It can
easily be misinterprted as suppressing all output from the command, but
that's not what it does. I may revisit.