Improve string lexing & escaping
I hadn't given the correct handling of escaping in string lexing enough
thought. It's been servicable so far, but in this commit, we revisit it
a bit, and handle escaping for \t and \n as well as the existing { .

I describe the rules around escaping a bit on a new Strings MkDocs page.
Basically, single and double quote strings have the same rules and allow
more escape characters. Backticks allows less but still has e.g.
interpolation, which I'm opting to keep because backticks make for good
shell command delimiters but also interpolation is useful when putting
those together.