Implement function 'parse_json'
Kinda like the Python JSON loads function. Takes in a string and tries
to parse it as json, converting it recursively into native RSL types.
For example, if you have a valid JSON dictionary as a string, and pass
that in, it will return an RslMap with the constituent fields (all the
way through) likewise converted to native RSL types.

This should be useful for operations where you get a JSON string returns
(e.g. via a shell command or a rad request) and you want to
inspect/interact with the json blob as it is.