Fix -h when used on scripts
Not sure how this slipped through for so long, but when -h or --help is
invoked on a RSL script, we're passing the wrong thing to the "read
source file" command, resulting in a "can't read this file" error, and
no help for the user.

Simple fix. Add a test to test -h.