Delete Cobra: Handle arguments ourselves using pflag
This has been a long time coming, and is a pretty big change. I
originally used Cobra because it gives you a lot of things and seemed
like the go-to for making Go cli tools. What I should probably have seen
coming was that our requirements for rad were too particular and needs
to be quite custom. If I were to try and implement all the behaviors I
wanted, I'd be fighting Cobra the whole time (we already see a little of
this with the double command execute), overriding a ton of pieces it's
meant to give me, defeating the purpose of using a library like this,
and it'd be overly complicated compared to just doing this linearly
ourselves.

So it's clear we should drop it, that's what we do in this commit. I had
to decide whether to keep pflag as well (which Cobra also uses) or
should go straight with vanilla Go flags, but I decided to keep pflags
as that really does give us a lot of nice things that the built-in flag
lib does not, without being too opinionated, still.

Theres' too much to go through in this commit but TLDR is the runner.go
contains most of the logic. We try parsing flags with pflag, make some
assessments based on what we read, we may then read a script if we think
there's RSL to read, add the custom args/flags, and re-parse. From
there, we again assess what to do and do that e.g. interpeting and
executing script.

A nice consequence of this is that the rad binary also gets reduced from
14092 KiB to 11220 KiB.