Replace global 'version' flag with VERSION
More consistent with other global flags.

I thought about removing the long name entirely and just having -V, so I
made some slight changes in the usage generation code to support that,
but ultimately decided against that for this flag.

Now our release script can safely have an arg named 'version' so we
rename the 'new_version' one to that.