Internal: replace 'string' with 'RslString' struct
Right now, strings are being passed around in their raw form. This can
be a *little* restrictive, as I have a few use cases for wanting to
store metadata / attributes for strings. This includes colors,
formatting such as bold/italics, whether a string should be considered a
regex or not, etc.

To be more specific, I'd like to allowed something like this:

  name = red("Alice")
  print(name)

I *could* make red() instantly apply escape chars onto "Alice" but this
makes further string manipulation like len(), substringing, etc, very
fraught. Color should only printing, basically. By introducing
RslString, we can store the color as an attribute separate from the
string itself, allowing operations to stay on the 'pure', unchanged
string.

The other use case is e.g. regex. Right now, we have a replace()
function which can look like this:

  replace(my_text, "1+", "1-")

One might think this is valid and replaces any '1+' substrings with
'1-', but replace() actually treats the second two strings as *regexes*,
so the '1+' actually gets interpreted as 'one or more 1s'. This can
understandably catch people out. And the solution is definitely not to
remove the regex support -- it's quite useful. I also don't love just
making regex versions of functions e.g. replace_regex(). I'd rather
allow users to somehow communicate 'this is a regex string, treat it as
such'. So in the future, we can do this:

  replace(my_text, "1+", "1-")  // plain text "1+" replacement
  replace(my_text, re"1+", "1-) // regex "1+" replacement

This is just an example, syntax is not locked in, but this could enable
a 'isRegex' metadata field on the RslString struct, which our replace()
implementation could then check in order to know whether or not the
string should be treated as a regex or plain text.

This commit simply sets up the infrastructure for this by replacing all
usages of 'string' as internal RSL string representations with an
RslString struct. A few of the implementations are very naive e.g. the
non-cached .Plain() implementation, but this stuff can be easily
optimized later.

Unfortunately our uniq() implementation needs to be a little custom now.
It doesn't look super trivial to make RslString meet the requirements
for using lo.Uniq, so we just implement our own 'ToPrintable'-based
solution. Maybe a bit naive as well, but it does the trick for now.

:: Extra ::

While here, we also disallow this:

a = [1, 2]
a += 3

Rather than adding 3 to the list, we'll just throw an error. This is
more in line with other languages and less error-prone.