Add string-coloring functions
The previous commit introduced RslStrings so that we could track
'attributes' for strings separately from their underlying text. In this
commit, we add the first attribute: color.

We allow seeing this attribute on strings via color functions. For
example:

  a = yellow("Alice") // returns a string "Alice" with color = yellow

'a' will behave as a normal string in (almost?) every way, except for
when printed, in which case we consult the attributes of the RslString,
and if a color is specified, we use fatih/color to colorize the string
before printing it out.

There are many colors allowed. As implemented, any color allowed under
the 'rad' block coloring functionality is also available as a function.
I debated whether or not I should expose them all as functions as
opposed to something like this:

  a = color("Alice", "red")

I've decided to go with the 'colors are functions' approach because RSL
is a scripting language and so we do put a fairly high premium on
writeability (especially when readability does not suffer).

We may offer more flexible functions in the future e.g.

  a = rgb("Alice", 128, 255, 255)

This commit does take some shortcuts, which I may patch up in the
future. I want to commit to all 'pure' string operations *maintaining*
string attributes. In case it's not clear, there's a bit of complexity
on what to do when operating on strings with attributes. For example:

  a = red("Hi ") + green("Alice")

RslString internally tracks 'segments' separately from one-another,
where the attributes are per-segment. So in this case, the concatenation
leads to an RslString with two segments internally, each tracking their
colors appropriately. This then properly gets printed as you'd hope if
print(a) gets invoked.

As I was saying, 'pure' string operations such as indexing, slicing,
concatenation, should maintain attributes. That's not the case in this
commit -- I've left slicing for later due to the complication where you
want to slice *across* segments. For example, a[1:5] in the above
example would be 'i Al' where 'i' is red, and 'Al' is green. The
implementation is not *hard*, I just need to get to it, and it's not in
this commit!

What I'm a little less committed to is proper handling for all string
functions. For example, replace() is probably the most complicated one.
You can imagine having a segment, replacing some text in the middle of
it, and now you've got three segments (left side, the middle-replaced
bit, and the right side). Again, not *hard*, I just need time to get to
it. But from an API/docs perspective, I've decided at least for now to
have the 'preservation of attributes' be at a per-function discretion. I
will *probably* do my best to implement it everywhere, but for now,
replace() and split() just return plain strings. upper() and lower() are
very easy though, so I've made them preserve attributes.

One straight-up bug right now is regarding equality and the unique()
function, demonstrated here:

  a = "Alice"
  b = red(a)
  print(a == b)          // prints true
  print(unique([a, b]))  // prints [Alice, Alice] (second in red)

Obviously, if a == b, then they should not both be present in the array
after unique() is called on it. This is happening right now because the
unique() function is implemented by running ToPrintable on all elements
and hashing those. The red Alice gets hashed differently from the plain
Alice, but for our RslString Equals function, we just compare the plain
strings.

I'm not 100% certain which way I want to take this. Should attributes be
considered in equality? I'm inclined to say no.