Fix bool args not defaulting to false
In my deletion of Cobra, I had inadvertently broken bool args defaulting
to false if not given an explicit default and not configured on
invocation. It would erroneously be detected as a 'missing arg' and
trigger a usage print -> exit.

Fix is simple - uncomment a line I thought was unnecessary before, and
add some test coverage.