Bug fix: support RslString for 'in' expressions
  a = ["alice", "bob", "charlie"]
  print("alice" in a)

this broke in a recent-ish commit where I replaced strings with
RslStrings, unfortunately did not have test coverage for strings with
'in' expressions.

Disappointingly in Go, you cannot define a method on structs which will
effectively override the '==' operator for that struct. The method doing
our comparison is this one:

  func contains(array []interface{}, val interface{}) bool {
    for _, v := range array {
      if v == val {
        return true
      }
    }
    return false
  }

so it'd be quite neat if we could just override == for RslString and
leave this method as-is, but this doesn't seem possible, so I need to
explicitly detect if the types I'm comparing are RslString, and if they
are, invoke a bespoke Equals method on the structs. Greatly surprising,
maybe I'm missing an idiomatic way to do this better in Go.

Anyway, that's what we do for this commit + add test coverage.