Allow named arg 'prompt' in pick and pick_kv
This was the plan from the start, basically customizing a text prompt
when selecting with these two functions. I also want to extend this to
pick_from_resource, but I've not yet implemented the fuzzy huh selection
logic for that one yet -- it will come.

This is sorta a two-for-one commit. This also fixes a longer-standing
'bug' or strange behavior that I've now tracked down to huh when you
give it an empty prompt/title. Basically, it always hides one of the
available options, and if the user starts to narrow down the options by
typing, it doesn't display what they're typing back to them.

Seems like undesirable behavior from huh. So we default to always having
a "Pick an option" prompt, and if the user tries to set an empty prompt,
rather than using the undesirable huh behavior, we make the title empty,
which looks a little odd too but at least the user gets to see what
they're typing and all the options.