Implement function 'http_get'
As usual, the current way to do this would be via a shell cmd, but let's
avoid delegating all these things to shells. If we can offer
high-quality utilities without needing to leave RSL in any way, that's a
good thing.

Here, we implement a simple http GET function. It takes a URL and
optionally has a named arg for an RslMap containing headers.

For a successful query, it returns a map with two keys: 'body' and
'status_code'. There're more things to add here, but that can be done
later.

Regarding the response -- if the response is a valid JSON blob we can
parse, we do that and convert it to native RSL types. If there's a use
case for just leaving json responses as strings, then in the future we
might add a optional bool to the http_get method to disable that.

If the query fails, we just error and exit the script. We might want to
offer some ability to handle errors here, but I'm not sure what form
that'd take. We could adopt some try-catch style logic (though I don't
love that), alternatively we offer a new syntax for something like this
method to return a value that must we unwrapped a la Rust, and this can
allow for error handling.

I noticed, while doing this change, that our pprint function does some
HTML escaping. I'm not convinced we want that, so I change
rsl_func_print.go a little bit to disable that in the jsoncolor
formatter.