Add .dot.key.access syntax for maps
Rather than having to access keys like this: myMap["key"]
You can now do this: myMap.key

Part of the motivation is that it's a little easier to type and read
(although the latter is not quite as compelling, given readers will now
need to be aware of both syntaxes, which is a downside).

The main motivation, however, is so that we can approximate 'structs',
with the immediate use case being able to return errors from operations
that may fail, and allowing users to handle that. For example, right now
you can do this:

  a = int(myNum)

If this succeeds, a is now an integer. However, if it fails, RSL will
just throw and the developer has no chance to program in a recovery
mechanism (we don't have try-except patterns).

I want to give devs an ability to handle this. However, I don't love the
try-catch approach. I'd rather take a more Go-inspired solution and
return an error object that can be checked. Here's what it could look
like:

  a, err = int(myNum)
  if err:
    print("Couldn't parse {myNum}! Error: {err.msg}")
    exit(1)

Note here the 'err.msg' syntax. My intention is to have int() be able to
return *two* outputs. First, the int (0 if failed), and secondly the
error. The error will be an RslMap under the hood. And by allowing the
dot.syntax in this commit, devs can treat this more like a struct. It's
a bit Python-esque, in that Python objects are also all just
dictionaries underneath.

Small other note on the above example: that if-statement implies
truthy/falsy logic. I also intend to implement that. If the parse was
successful, err will just be an empty dictionary, which will be falsy.

==  Implementation  ==

In terms of implementation, I basically treat this dot.syntax as
syntactic sugar. When parsing, it becomes the exact same instruction as
if you had done ["this"] syntax.

The intention is also to allow nested assignment e.g.

  person.address.country = "Australia"

but we don't have the logic for this nested approach yet (it does not
exist for the 'standard' syntax either). Soon to come, hopefully.

Small note re: the removal of 'ColonToken' from 'CollectionAccess' is
a simplification, in part because we don't actually use it anyway,
either.