Add truthy/falsy logic
Truthy/falsy is a pretty nice concept for script brevity and (once
you're familiar with it) readability. Given the prominence of Python and
truthy/falsy logic in Python, I think it's acceptable to add in RSL,
especially because of the scripting use case calling for brevity.

The implementation is pretty straight-forward and hopefully
unsurprising. The following values are falsy, everything else is truthy:

| Type   | Falsy |
|--------|-------|
| string | ""    |
| int    | 0     |
| float  | 0.0   |
| list   | []    |
| map    | {}    |