Rename 'int' and 'float' funcs to 'parse_int' and 'parse_float'
The initial naming was very concise, which was the appeal, and also
close it Python. However, on reflection, I think it'll be better to
prefix if 'parse_' to make more explicit that these are *parsing*
functions, not *casting* functions. Other langs e.g. Go/Zig/Rust also
all use the word 'parse' in some capacity, so it won't be terribly
surprising for users and those are probably good languages to take
examples from.