parse_int, parse_float: Allow reading error map to avoid script exit
The past few commits have been building towards this kind of construct.
Right now you can do this:

  a = parse_int(myNum)

If this succeeds, 'a' is now an integer. However, if it fails, RSL will
just throw and the developer has no chance to program in a recovery
mechanism (we don't have try-except patterns).

This commit is a move towards giving devs an ability to handle this, and
sets a general direction for how we want to do these things. I don't
love the try-catch approach some languages take, and I'd rather take a
more Go-inspired solution and return an error object that can be
checked. Here's what it looks like after this commit:

  a, err = parse_int(myNum)
  if err:
    print("Couldn't parse {myNum}! Error: {err.msg}")
    exit(1)

Bit of an obtuse example since we exit anyway, but you could imagine
e.g. retrying user input, etc.

If the above parsing succeeds, 'a' will be the int and 'err' will be an
empty map. Thanks to the recently-added truthy/falsy logic, the if
statement will not run.

If the parsing fails tho, 'a' will just be 0 (generally values should
return a falsy value of the expected type, probably), but the 'err' map
will look something like this:

  {
    code: "RAD20001",
    msg: "Failed to parse 'asd' as an integer"
  }

This will be truthy, hence the if-block will execute.

Note the code. This is still very early but I might commit to this
generally for Rad/RSL errors. It's inspired by TypeScript having their
own 'TS'-prefixed error codes, which make searching up errors and
solutions for them a bit easier.