Improve URL encoding: spaces, percents
I had missed that my existing implementation was *very* basic and didn't
even encode spaces. Now it does. Also, % should be encoded as %25,
otherwise it will fail. We simply do a find-replace as it appears tricky
to get Go's tooling to do this for us.

This seems to do the trick, I'm sure I'll hit more cases that don't, but
will fix those as we go.