Replace collection/slice access with VarPath
The AST contains some types that we can replace if we alter how VarPath
works. That what this change does. This *should* help simplify an
upcoming change I'm trying to implement which would allow e.g.

  myMap["key"][2] = 2

to be possible i.e. nested-key entry-assignment.

I think we can collapse some more types, but this change is made a
little hesitantly as I'm not 100% sure that this path will work. This is
the first step to finding out though -- giving it a try.

In implementing this, we technically have a new feature, which is
deleting slices of characters from strings:

  a = "alice"
  del a[1:3]
  print(a)  // prints "ace"

Am also relying on test coverage quite a lot here. It's all passing, but
will keep an eye out for bugs introduced by this where there're gaps.