Replace CompoundAssign AST with basic Assign
You can think of this as us basically 'pushing up' the point where we
begin treating compound assignment as simply a variation of a binary
operation with an assign. This allows the two assignment types to share
paths earlier, allowing for less code downstream handling them.

This does involving rejigging the Binary struct a little, to separate
the Token we use for errors from the actual operation we want to
execute when the Binary is being interpreted.