Replace CollectionEntryAssign with 'Assign' AST type
The last few commits have been building towards this. Right now,
'collection entry assignment' is limited to one lookup, but we want to
support arbitrary levels of lookups, which we can accomplish via
VarPath. In other words, with this commit,

  myMap.key[1] = 2

is now allowed. Not only that, but if you have a function returning two
values e.g. pick_from_resource, you can do collection entry assignment
with that too:

  a[1], a[2][1] = pick_from_resource(...)

Not itself a selling point or goal, but a consequence of the
implementation now being better able to handle arbitraty var paths.

One downside is that, by making the handling for Assign more generic and
apply to larger variety of contexts, our errors get a bit less
descriptive. Am thinking we could inject some more context depending on
where we create the instructions in the parser.

A couple of things still remain in the 'var path' overhaul of commits
I've been working on:

- switch assignments
- shell cmd output assignment