Add arg constraint syntax: string enum
  args:
    name string
    enum name ["alice", "bob"]

With this syntax, you can now declare *within an arg block* an
enumerable list of valid values (hence the 'enum' keyword).

The syntax might be a little controversial -- it's not ideal that we
need to repeat 'name' and write the word 'enum'. You could conceive of a
more concise way to implement this, but we don't go that route for two
reasons:

1) The arg declaration lines are already quite dense. To remind readers:

  varName "alias" shorthand type = default # comment

This is the current anatomy of what's allowed on an arg declaration. I
don't love the idea of complicating this more by adding this 'enum' use
case in for string arguments.

2) Related to the first, but 'enum' is not the only constraint I want to
   implement, and so building constraints into the arg declaration
   itself won't scale well.

You could concieve of a compromise where you do something like:

  args:
    name string:
      enum ["alice", "bob]

but I'm not convinced this is that much better. You only avoid the
'name' repeat, but at a cost:

1) a colon and an indentation
2) loss of flexibility for where you can declare that enum line

To think a bit forward and demonstrate why I think this direction in arg
constraint syntax will be better, take this example:

 args:
   command c string # Which command to execute.
   name string = "Alice" # The name of the person.
   value int = 10 # The value to run with.
   debug bool # Enable for debug output.

   enum command ["run", "query", "kill"]
   regex name "^[A-Z].*"
   range value [0, 20)

This shows an additional couple of constraints I could concieve making
(regex and range). I wouldn't want to begin squeezing these things into
already-dense arg declaration lines.