Print one-liner description in usage if no block description
Still need to revisit this whole 'one liner' and 'block description'
split, but right now, if you specify e.g.

   ---
   some description
   ---

in a script, then this description doesn't get printed at all because it
gets put into the 'one liner' field, and we leave the block empty, but
in our usage logic, we only look at the block currently.

In this commit, we fall back to the one-liner if it's there and the
block description is not.