Parser bug fix: try consume new lines after indent in blocks
Prior to this commit, the following would fail:

  if true:
    // comment
    print('hi')

because the parser is handed an extra newline token due to the comment
after the initial 'INDENT' token. We try to read one of the lines as a
statement but it's just a newline, so fails because it expects
*something*, like an expression.

Let's just add another consumeNewLines call.