Fix padding for colored strings: account for ASCII characters
Currently, padding a colored string doesn't work correctly.

  "{name:20}"
  "{blue(name):20}"

visually, you'd expect identical padding between these two. The only
differences should be the color. However, in the colored case, the
padding logic is actually dealing with a string surrounded by ASCII
color escape characters. These characters consume the padding, and so
the blue string appears to have less padding.

This is undesirable, and we correct for that in this commit.

Unfortunately it does mean complicating the formatting-related code a
little. When we are parsing the user's format, we extract out the width
into an explicit int64 we'll save for later, and instead of appending
the width as-written into the format string, we append special 'NULL'
chars, that we will find & replace later.

That's done in the parser. When we finally are formatting in the
interpreter, we detect if the item being formatted is a string. If it
is, we compare it's "plain" length to the proper formatted string we're
gonna ask Go to format. The plain string is presumably equal or shorter.
Any difference is added to the width we saved earlier from the parsing
stage, and then this final width is what we use for the formatting.

If there's no width specified or the item is not a string, none of this
logic happens.