Loosen variable visibility rules after blocks
  if true:
    a = 1
  print(a)

Prior to this commit, the above would fail at runtime because the
variable 'a' gets "forgotten" once we exit the if-block, since it was
defined in there.

Similarly,

  for i, item in ["alice", "bob"]:
    ...
  print(i, item)

would also fail because variables i and item are forgotten.

I had initially implemented these 'stringent' rules since that's how a
lot of C-like languages operate, including Java.

However, I've since changed my mind on this with regards to RSL due to
the focus being on scripting. When you're scripting, having more relaxed
rules can be more productive, and having strict rules around correctness
can just get in the way and make the scripts more cumbersome to write.

I wasn't actually sure of this before but Python actually also behaves
this way. Variables defined in blocks, like the examples above, are
still accessible after the block ends. So we're not breaking new ground
here with RSL, and instead stay consistent with what e.g. Python users
may expect.

None of this applies to lambdas in rad-block mappnings:

  rad url:
    fields nums
    nums:
      map n -> n * 10

in the above example, I don't think we want the 'n' to remain an
accessible variable after the rad block has completed, and if 'n' is
already defined prior, I don't think we want it modified to the last
value that 'n' takes in the lambda iteration.
I might be proven wrong in this respect, and we can relax the rules here
too, but for now, let's keep the lid on it and let the lambda run in
isolation.