Use own pflag flagset
Prior to this commit, we're using the pflag 'singleton' CommandLine
flagset. This flagset is instantiated with ExitOnError, but we want
ContinueOnError. The only way to achieve this is instantiate our own
flagset, so let's do that.

This command simply replaces our use of the CommandLine singleton with
our own singleton. I'll follow up with commits to actually leverage this
for rad improvements.