Implement raw strings
I've been giving strings a lot of thought. Right now, RSL is missing the
ability to easily encode regexes in strings such as [0-9]{2}, because
the {2} gets interpreted as an expression (string interpolation) and you
therefore end up with a string "[0-9]2". You could escape the { char,
but now we're obfuscating the regex.

Backticks have different escaping rules from single and double quotes,
but they still have string interpolation. The reason for this was
originally because backticks is the go-to for writing shell commands,
and it's useful to have interpolation when writing shell commands.

This commits changes things up a bit. We make backticks behave the same
as single and double quotes, and we add Python-style 'r-strings' to RSL.

The syntax is simple: prefix whatever delimiter you want to use with
'r', and you'll end up with a raw string. Raw strings treat all the
contents as-is, meaning something like \n will literally be treated as a
backslash followed by an n. It also ignores string interpolations, so
they too remain as-is.

You can still escape the string delimiter itself, but that's it.

Example:

  print(r"hi\n{name}")

This will get printed literally as it appears above. One line, no
interpolation.

Prior to this change, I posted on /r/ProgrammingLanguages to solicit
input from the folks there, link here:

  https://www.reddit.com/r/ProgrammingLanguages/comments/1homp3s/

Asides from this raw string change in this commit, I'll probably be
making a couple of more changes:

1) Make interpolating use ${} instead. I am convinced the simple {}
   syntax (especially given it's on by default) could be annoying if
   you're trying to e.g. paste JSON into a string. ${} is quite common
   as well, only Kotlin does away with the $, as far as I can tell.

2) Implement """ Python-style multiline strings.

Heredocs and custom delimiters a la Ruby/PHP/Perl are very interesting,
but tbh I've very rarely been cornered so badly that I need my own
delimiters, and I can't think of a straight-forward syntax that is
readable to someone unfamiliar with any of these concepts or RSL, as is
a guiding principle of the language.

With regards to the actual issue of representing regexes, I've not yet
made my mind up entirely. With raw strings, we have a few more options.

One might simply be to pass a raw string into a re() function, and the
returned value is a string with isRegex set, meaning you can do e.g.

  out = replace(input, re(r"[0-9]{2}"), 10)

though this is also a bit of an eyesore. We could do

 out = replace(input, /[0-9]{2}/, 10)

i.e. use /regex/ to express regexes, which some other languages do. This
wouldn't leverage raw strings but we'd probably re-use some of the raw
string parsing logic for it.

Anyway I'm quite off-topic from this actual commit, will think more.